% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate.R
\name{collate}
\alias{collate}
\title{Collate data from different water/sediment BLE LTER Core Program data sets}
\usage{
collate(
  ids = NULL,
  output = "object",
  path = NULL,
  avg_rep = FALSE,
  skip_metadata = FALSE
)
}
\arguments{
\item{ids}{(numeric) Vector of BLE dataset IDs to grab. Default is to get these dataset IDs: 2, 3, 4, 11, 12, 13, 18. To quickly reference which IDs correspond to what data, run \(code){which_id()} in the R console.}

\item{output}{(character) Choice of "excel", "csv", or "object". Returns one Excel file or many CSV files in addition to a R list of data frames, or skipping writing to file altogether, respectively. Defaults to "object".}

\item{path}{(character) Path to working directory. Data files will be written to this directory. Defaults to the R session's working directory if unspecfied.}

\item{avg_rep}{(logical) TRUE/FALSE on whether to average replicates. This really only affects the nutrients dataset (knb-lter-ble.14) because this is the only dataset in consideration still retaining replicates in the published version. If FALSE, any data with replicates will be pivoted to a wider format, with the rep number appended to the new column names. E.g., two rows (reps 1 and 2) of one column "ammonium_umol_N_L" become one row of two columns "ammonium_umol_N_L_rep1" and "ammonium_umol_N_L_rep2". If TRUE, numeric columns will be averaged (NAs are ignored) and character columns will be collapsed into one string (e.g. if two replicates from the same sample have the flags VALID and BD, this becomes "VALID BD"). Note that in the original nutrients data, 2018-2019 reps are always NA, because we did not report replicates for these years. "NA" reps become rep 1 for the purposes of this package. Defaults to FALSE.}

\item{skip_metadata}{(logical) whether to skip metadata, defaults to FALSE}
}
\value{
(list) Named list of data frames: "water_data", "sediment_data", "dataset_metadata", and "column_metadata"
}
\description{
One stop shop for your data collation needs. This function, using defaults with no parameters or arguments, will query the Environmental Data Initiative (EDI) repository for the latest versions, downloads those, and collates data and metadata. See the available arguments for available customization options.
}
