% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_data.R
\name{collate_data}
\alias{collate_data}
\title{Collate data from different water/sediment BLE LTER data sets}
\usage{
collate_data(data, output = "excel")
}
\arguments{
\item{data}{(list) Nested list of data frames, output from download_data}

\item{output}{(character) Choice of "excel", "csv", or "object". Returns one Excel file, many CSV files, or an in-memory list, respectively. Defaults to "excel".}
}
\value{

}
\description{

}
