% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate.R
\name{collate_metadata}
\alias{collate_metadata}
\title{Summarize BLE dataset and column metadata}
\usage{
collate_metadata(ids = NULL, metadata)
}
\arguments{
\item{ids}{(integer) Inherited from the function call to collate}

\item{metadata}{(list) List of metadata}
}
\value{
(list) List of two data.frames: "dataset_metadata" and "column_metadata"
}
\description{
Summarize BLE dataset and column metadata
}
