% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download BLE data files}
\usage{
download_data(ids = NULL, path = NULL, write = F)
}
\arguments{
\item{ids}{(numeric) Vector of dataset IDs to grab}

\item{path}{(character) Path to write CSVs to}

\item{write}{(boolean) Whether to write to file}
}
\value{
(list) A nested list of data frames. 1st-level list item corresponds to dataset, 2nd-level items correspond to entities within a dataset and contain a data.frame of the data entity. List items are named accordingly (dataset after the packageId and entities after the entity name). Optionally, if write=T, functions also write to CSVs files in specified path. File names are the full package Id, followed by two underscores, followed by the full entity name.
}
\description{
Download BLE data files
}
