% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_uncollated_dfs}
\alias{get_uncollated_dfs}
\title{Download BLE data files slated for collation}
\usage{
get_uncollated_dfs(ids = NULL, path = NULL, write = FALSE, metadata = TRUE)
}
\arguments{
\item{ids}{(numeric) Vector of dataset IDs to grab, default is to get all data}

\item{path}{(character) Path to write CSVs to, defaults to working directory}

\item{write}{(boolean) Whether to write data entities to file, defaults to FALSE}

\item{metadata}{(boolean) whether to get metadata, defaults to TRUE}
}
\value{
(list) A nested list of data frames. 1st-level list items are named "data" and "metadata" (if metadata = TRUE). For "data": 2nd-level list items correspond to datasets, 3rd-level items correspond to entities within a dataset and contain a data.frame of the data entity. List items are named accordingly (dataset after the packageId and entities after the entity name). Optionally, if write=T, functions also write to CSVs files in specified path. File names are the full package Id, followed by two underscores, followed by the full entity name. For "metadata": 2nd-level list items correspond to datasets, and contain a the metadata for that dataset in "emld" aka list format.
}
\description{
Download BLE data files slated for collation
}
