% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_uncollated_entities}
\alias{get_uncollated_entities}
\title{Get entity names and IDs from newest revisions of BLE data sets slated for collations}
\usage{
get_uncollated_entities(ids = NULL)
}
\arguments{
\item{ids}{(numeric) Vector of BLE dataset IDs to grab}
}
\value{
(list) A nested list. 1st-level list item corresponds to dataset, 2nd-level items correspond to entities within a dataset and contain a string of the entity Id. List items are named accordingly (dataset after the packageId and entities after the entity name).
}
\description{
Get entity names and IDs from newest revisions of BLE data sets slated for collations
}
