% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess tables before putting them info the join}
\usage{
preprocess(df, ysi = F)
}
\arguments{
\item{df}{(data.frame) ONE data.frame containing BLE LTER water/sediment core data, as downloaded from EDI without any processing.}

\item{ysi}{(logical) Set this to TRUE when the data is YSI. Returns date_collected as usual, plus a date_time_YSI column that retains the time.}
}
\value{
(data.frame) A data.frame
}
\description{
This (1) renames any date_time column to date_collected, and remove the time portion; (2) renames any occurrences of "STL" into "SSL"
}
