% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ble_packages.R
\name{search_ble_packages}
\alias{search_ble_packages}
\title{Search all BLE LTER packages}
\usage{
search_ble_packages(fields = "id,packageid,title,pubdate", env = "production")
}
\arguments{
\item{fields}{(string) Searchable metadata fields in EDI repository to include query that produces output.
The way this function is written now, this string must follow the Solr query syntex (e.g. "id, packageid, title").}

\item{env}{(string) Repository environment. Can be "production", "staging", or "development".}
}
\value{

}
\description{
Modify EDI utils search_data_packages function to return only BLE LTER data. Default is to provide a dataframe with
all BLE package id ("scope.identifier.revision" format), title, publication dates, and the package number from
package id.
}
\examples{

}
