% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_data.R
\name{collate}
\alias{collate}
\title{Collate data from different water/sediment BLE LTER data sets}
\usage{
collate(
  ids = NULL,
  output = "object",
  path = NULL,
  avg_rep = FALSE,
  skip_metadata = FALSE
)
}
\arguments{
\item{ids}{(numeric) Vector of BLE dataset IDs to grab. Default is to get all data.}

\item{output}{(character) Choice of "excel", "csv", or "object". Returns one Excel file or many CSV files in addition to a R list of data frames, or skipping writing to file altogether, respectively. Defaults to "object".}

\item{path}{(character) Path to working directory. Data files will be written to this directory. Defaults to the R session's working directory if unspecfied.}

\item{avg_rep}{(logical) TRUE/FALSE on whether to average replicates. Defaults to FALSE.}

\item{skip_metadata}{(logical) whether to skip metadata, defaults to FALSE}
}
\value{
(list) Named list of data frames: "water_data", "sediment_data", "dataset_metadata", and "column_metadata"
}
\description{

}
