% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_data.R
\name{collate_data}
\alias{collate_data}
\title{Collate data from different water/sediment BLE LTER data sets}
\usage{
collate_data(ids = NULL, output = "object", path = NULL, avg_rep = F)
}
\arguments{
\item{ids}{(numeric) Vector of dataset IDs to grab, default is to get all data}

\item{output}{(character) Choice of "excel", "csv", or "object". Returns one Excel file, many CSV files, or an in-memory R list of data frames, respectively. Defaults to "object".}
}
\value{

}
\description{

}
