% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R, R/get_data.R
\name{download_data}
\alias{download_data}
\title{Title}
\usage{
download_data(ids = NULL, path = NULL, write = F)

download_data(ids = NULL, path = NULL, write = F)
}
\arguments{
\item{ids}{(numeric) Vector of dataset IDs to grab}

\item{path}{(character) Path to write CSVs to}

\item{write}{(boolean) Whether to write to file}
}
\value{


(list) A list of data frames. Optionally, CSVs files in specified path
}
\description{
Title

Download BLE data files
}
